// src/api/models/team-invitation.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Your API
 * API description
 *
 * The version of the OpenAPI document: v1
 * Contact: dev@example.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface TeamInvitation
 */
export interface TeamInvitation {
    /**
     * 
     * @type {number}
     * @memberof TeamInvitation
     */
    'id'?: number;
    /**
     * 
     * @type {number}
     * @memberof TeamInvitation
     */
    'team'?: number;
    /**
     * 
     * @type {string}
     * @memberof TeamInvitation
     */
    'email'?: string;
    /**
     * 
     * @type {number}
     * @memberof TeamInvitation
     */
    'invited_by'?: number;
    /**
     * 
     * @type {string}
     * @memberof TeamInvitation
     */
    'invited_by_email'?: string;
    /**
     * 
     * @type {number}
     * @memberof TeamInvitation
     */
    'user'?: number | null;
    /**
     * 
     * @type {string}
     * @memberof TeamInvitation
     */
    'user_email'?: string;
    /**
     * 
     * @type {string}
     * @memberof TeamInvitation
     */
    'team_name'?: string;
    /**
     * 
     * @type {string}
     * @memberof TeamInvitation
     */
    'status'?: TeamInvitationStatusEnum;
    /**
     * 
     * @type {string}
     * @memberof TeamInvitation
     */
    'token'?: string;
    /**
     * 
     * @type {string}
     * @memberof TeamInvitation
     */
    'created_at'?: string;
    /**
     * 
     * @type {string}
     * @memberof TeamInvitation
     */
    'updated_at'?: string;
}

export const TeamInvitationStatusEnum = {
    Pending: 'pending',
    Accepted: 'accepted',
    Rejected: 'rejected'
} as const;

export type TeamInvitationStatusEnum = typeof TeamInvitationStatusEnum[keyof typeof TeamInvitationStatusEnum];